arch/arm: fix convert conditional inst
Before a block of instructions (list) was replaced by
a new branch which will contains the splited block (list
of instructions and new Ast_If_cond).

Now all new nodes are not moved in a new branch, but replaced
in the current ast : the block is removed, and new nodes are
added. It fixes some fusions of cmp - Ast_if_cond (without
the modification we needed to look inside the branch, and again
in the branch of Ast_If_cond).