fix #45: add equivalent at same deep on loop dependencies tree
Two loops are equivalent if all addresses of the loop and sub-loops
are equals, and the test is done only if the two loops are at the same
level.

Note: before the definition of equivalent loops was a bit different :
Two loops were equivalent if all addresses were equals (but not the sub-loops).